pyproject.toml
[project]
name = "advent-of-code-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
create = "uv run advent_of_code/create.py"
day = "uv run advent_of_code/run.py"
